requirements.txt
streamlit
openai
numpy
pandas
matplotlib
scipy
sounddevice
soundfile
python-dotenv
gtts
streamlit-lottie
fpdf
requests